implement separate brightness controll to allow fade in&out via brightness
When reducing the brightness via Neopixel library (Adafruit, NeopixelBus, etc.) the original color will not be retained. So after setting brightness to 1 and afterwards back to 255, the color range is lost compared to the original color settings.

Therefore the array "current_colors" holds the original colors so brightness changes does not kill the original color settings.